requirements.txt
keras-gpu==2.2.4
h5py==2.10.0